requirements.txt
streamlit
psycopg2-binary
python-dotenv
passlib
python-jose
watchdog
traceback